+++ esm/index.js
@@ -5,7 +5,7 @@ import DataBuffer from '../src/data-buffer.js';
 import DataBufferList from '../src/data-buffer-list.js';
 import DataStream from '../src/data-stream.js';
 import LZW from '../src/data-compression-lzw.js';
-import { formatBytes } from '../src/data-formating.js';
+import { formatBytes, hexTable } from '../src/data-formating.js';
 
 export default {
   CRC32,
@@ -15,6 +15,7 @@ export default {
   DataStream,
   LZW,
   formatBytes,
+  hexTable,
 };
 
 export { default as CRC32 } from '../src/data-hash-crc32.js';
@@ -23,4 +24,4 @@ export { default as DataBuffer } from '../src/data-buffer.js';
 export { default as DataBufferList } from '../src/data-buffer-list.js';
 export { default as DataStream } from '../src/data-stream.js';
 export { default as LZW } from '../src/data-compression-lzw.js';
-export { formatBytes } from '../src/data-formating.js';
+export { formatBytes, hexTable } from '../src/data-formating.js';