+++ test/tree-shaking/3-of-2.mjs
@@ -1,5 +1,4 @@
 /* eslint-disable import/extensions */
-/* eslint-disable unicorn/import-index */
 import { CRC32, LZW } from '../../esm/index.js';
 
 const main = () => {