+++ test/tree-shaking/4-of-1.mjs
@@ -1,5 +1,4 @@
 /* eslint-disable import/extensions */
-/* eslint-disable unicorn/import-index */
 import { DataBitstream } from '../../esm/index.js';
 
 const main = () => {